EnvelopeCollection is now abstract and create a TransportEnvelope
