+++ src/Transport/TransportEnvelope.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Zenstruck\Messenger\Test\Transport;
+
+use Symfony\Component\Messenger\Envelope;
+use Zenstruck\Messenger\Test\EnvelopeCollection;
+
+final class TransportEnvelope extends EnvelopeCollection
+{
+    private TestTransport $transport;
+
+    public function __construct(TestTransport $transport, Envelope ...$envelopes)
+    {
+        $this->transport = $transport;
+        parent::__construct(...$envelopes);
+    }
+
+    public function back(): TestTransport
+    {
+        return $this->transport;
+    }
+}
